requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-import-export==3.2.0
django-jsonfield==1.4.1
django-jsonfield-backport==1.0.5
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupPy==1.14
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
platformdirs==3.5.1
polars==0.18.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.4.0
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
